Fixed batch renderer issue

+++ Game/assets/shaders/batch.vert
@@ -2,10 +2,10 @@
 
 layout (location = 0) in vec3 a_Position;
 layout (location = 1) in vec2 a_TexCoord;
-layout (location = 2) in unsigned int a_TexUnit;
+layout (location = 2) in float a_TexUnit;
 
 out vec2 v_TexCoord;
-flat out unsigned int v_TexUnit;
+flat out uint v_TexUnit;
 
 uniform mat4 u_Proj;
 
@@ -13,5 +13,5 @@ void main()
 {
 	gl_Position = u_Proj * vec4(a_Position, 1.0);
 	v_TexCoord = a_TexCoord;
-	v_TexUnit = a_TexUnit;
+	v_TexUnit = uint(a_TexUnit);
 }
